+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/model/BaseEntity.java
@@ -4,6 +4,7 @@ import lombok.*;
 
 import javax.persistence.*;
 import java.io.Serializable;
+import java.util.Objects;
 
 @Getter
 @Setter
@@ -18,5 +19,22 @@ public class BaseEntity implements Serializable {
 
     public boolean isNew() {
         return id == null;
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        if (this == other) {
+            return true;
+        }
+        if (!(other instanceof BaseEntity)) {
+            return false;
+        }
+        BaseEntity that = (BaseEntity) other;
+        return Objects.equals(id, that.id);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id);
     }
 }